third_party/blink/web_tests/platform/generic/external/wpt/css/css-flexbox/flex-item-compressible-001-expected.txt

This is a testharness.js-based test.
PASS .flexbox 1
PASS .flexbox 2
PASS .flexbox 3
PASS .flexbox 4
PASS .flexbox 5
PASS .flexbox 6
PASS .flexbox 7
PASS .flexbox 8
PASS .flexbox 9
PASS .flexbox 10
FAIL .flexbox 11 assert_equals: 
<div class="flexbox">
      <div class="spacer"></div>
      <input type="text" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 100
FAIL .flexbox 12 assert_equals: 
<div class="flexbox">
      <div class="spacer"></div>
      <input type="range" class="test3" data-expected-width="140">
    </div>
width expected 140 but got 100
PASS .flexbox 13
PASS .flexbox 14
PASS .flexbox 15
Harness: the test ran to completion.
